requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
drf-base64==2.0
drf-extra-fields==3.7.0
filetype==1.2.0
pdf2image==1.16.3
Pillow==10.0.0
psycopg2-binary==2.9.7
PyPDF2==3.0.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
